Fix frame time

+++ Source/Runtime/Public/Components/ComParticleSystem.h
@@ -70,7 +70,7 @@ protected:
     virtual void            UpdateVisuals() override;
     void                    ChangeParticleSystem(const Guid &particleSystemGuid);
     void                    InitializeParticle(Particle *particle, const ParticleSystem::Stage *stage, float inCycleFraction) const;
-    void                    ProcessTrail(Particle *particle, const ParticleSystem::Stage *stage, float genTimeDelta);
+    void                    ProcessTrail(Particle *particle, const ParticleSystem::Stage *stage, int genTimeDelta);
     void                    ComputeTrailPositionFromCustomPath(const ParticleSystem::CustomPathModule &customPathModule, const Particle *particle, float t, Particle::Trail *trail) const;
     void                    ParticleSystemReloaded();
     void                    TransformUpdated(const ComTransform *transform);
